Code check

+++ frameworks/native/audioutils/src/volume_ramp.cpp
@@ -39,7 +39,7 @@ void VolumeRamp::SetVolumeCurve(vector<float> &volumes)
     CHECK_AND_RETURN_LOG(volumes.size() == VOLUME_SIZE, "Array size must 2!");
 
     curvePoints_.clear();
-    for (int32_t i = 0; i < times.size(); i++) {
+    for (size_t i = 0; i < times.size(); i++) {
         curvePoints_.emplace(times[i], volumes[i]);
     }
 }
